requirements.txt
numpy
tensorflow-gpu>=1.15.2
nltk>=3.4.5
